requirements.txt
Flask
flask-cors
langchain
langchain-groq
python-dotenv
mysql-connector-python
